requirements.txt
pandas
pandas_ta
click
mplfinance
matplotlib
pyyaml
colorama